Created new single kml file with all properties
Old kml files had one property with two separate
polygons (Burnet-2022-16) because it is split by a road. I made a merged
polygon for it that covered the road, then saved that in caltopo.
Unfortunately, that results in a kml file with points and polygons so I
cleaned that up by hand.